✅ Update tests for http_requests.py
